UX: Change NON_RESPONSIVE state to UNKNOWN
Its not necessarily that the application is non-responsive when the
background process fails to fetch its info. It could be the yarn proxy
is having issues, yarnitor itself cannot reach the proxy, etc. The
non-responsive state leads users to always think there's something wrong
with the app when really all it means is that the state of the app is
currently unknown.